move auth0 helpers to the new auth0service file
